requirements.txt
toml>=0.10.2
python-dateutil>=2.8.2
gitpython>=3.1.31

# Testing dependencies
pytest>=8.0.0
freezegun>=1.4.0 